Fix format warnings and add IO flow tier unit tests
- Fix format specifiers in integration/unit tests and eBPF loader
  * Use %zu for size_t in test_trans_integration.c
  * Use %lu with cast for uint64_t in test_trans_udp.c
  * Use %u for __u32 in dap_io_flow_ebpf.c

- Implement TODO functions in dap_worker.c
  * dap_worker_add_events_socket_inter: migrate to queue-based API
  * dap_worker_exec_callback_inter: migrate to queue-based API
  * dap_worker_context_callback_stopped: add proper queue cleanup

- Add comprehensive unit test for IO flow load balancing tiers
  * test_io_flow_tiers.c: tests tier detection, packet distribution,
    Application tier forwarding, and BPF response routing
  * Full mocking of system calls and worker functions
  * Tests designed to isolate and reproduce multi-client issues
  * All 4 test cases passing

- Update test build configuration
  * Uncommented io tests in tests/unit/CMakeLists.txt
  * Configured test_io_flow_tiers with proper includes and mocking

+++ tests/unit/net/trans/test_trans_udp.c
@@ -1460,8 +1460,8 @@ static void test_17_replay_protection(void)
         uint64_t l_sess_id = l_hdr.session_id;
         
         TEST_ASSERT(l_seq_num == l_expected_seq_nums[i],
-                    "Packet %d seq_num should be %u (got %u)",
-                    i + 1, l_expected_seq_nums[i], l_seq_num);
+                    "Packet %d seq_num should be %lu (got %lu)",
+                    i + 1, (unsigned long)l_expected_seq_nums[i], (unsigned long)l_seq_num);
         
         DAP_DELETE(l_decrypted);
     }
